requirements.txt
# Core dependencies
requests>=2.28.0
requests[socks]>=2.28.0
beautifulsoup4>=4.11.0
colorama>=0.4.6
lxml>=4.9.0

# Web scraping & automation
selenium>=4.10.0

# Network & Security
paramiko>=3.0.0
pwntools>=4.10.0
shodan>=1.28.0
dnspython>=2.3.0
python-nmap>=0.0.1

# Utilities
urllib3>=1.26.0
click>=8.1.0
tabulate>=0.9.0
pyyaml>=6.0

# OSINT & Information Gathering
maigret
python-whois>=0.8.0
whois
waybackpy>=3.0.6

# Web Analysis
python-Wappalyzer>=0.3.1

# Report Generation
jinja2>=3.1.0
markdown>=3.4.0
reportlab>=4.0.0
pillow>=9.5.0

# Environment
python-dotenv>=1.0.0